move platform-info
