+++ src/electron/window-management/platform-info.ts
